tweak: If not needed, don't do `--zip -e` to remove past zip.

+++ build.sh
@@ -76,7 +76,7 @@ if [ "$1" == "--runtime" ]; then
     doJob "Compiling runtime" "gcc -Idist/ runtime/voxel.c -o runtime/build/voxel"
 
     if [ "$2" == "--zip" ]; then
-        if [ -f runtime/build.zip ]; then
+        if [ "$3" == "-e" ] && [ -f runtime/build.zip ]; then
             rm runtime/build.zip
             echo "Removed old build.zip"
         fi
